Flytter DialectInput til egen komponent
