+++ app/routes/term.$termId.tsx
@@ -9,8 +9,7 @@ import { IconButton } from '@mui/material';
 import { IosShare } from '@mui/icons-material';
 import { ClientOnly } from 'remix-utils/client-only';
 import { useToggle } from '~/src/utils/use-toggle';
-import type { DictionaryResponse } from '~/routes/api.ordbokene';
-import { useDebounceFetcher } from 'remix-utils/use-debounce-fetcher';
+import { DialectInput } from '~/lib/components/dialect-input';
 
 export default function Term() {
   const { terms } = useRouteLoaderData<typeof rootLoader>('root');
@@ -210,52 +209,3 @@ export const ToggleButton = ({ leftLabel, rightLabel, name, handleChange }: Togg
     </label>
   </div>
 );
-
-function DialectInput({ name, dialect }: { name: string; dialect: 'nb' | 'nn' }) {
-  const fetcher = useDebounceFetcher<DictionaryResponse>();
-
-  function submitTerm(term: string) {
-    const formData = new FormData();
-    formData.append('term', term);
-    formData.append('dialect', dialect);
-
-    fetcher.submit(formData, { method: 'post', action: '/api/ordbokene', debounceTimeout: 200 });
-  }
-
-  function handleSuggestionClick(term: string | undefined) {
-    if (term !== undefined) {
-      const inputField = document.getElementById(name) as HTMLInputElement | null;
-      if (inputField) {
-        inputField.value = term;
-        submitTerm(term);
-      }
-    }
-  }
-
-  return (
-    <>
-      <input
-        id={name}
-        name={name}
-        className={'form-control' + (fetcher.data?.isValid ? ' is-valid' : '')}
-        type="text"
-        autoCapitalize="none"
-        onChange={(event) => submitTerm(event.target.value)}
-      />
-      <div className="valid-feedback bright-feedback-text">{fetcher.data?.validationText}</div>
-      {fetcher.data?.suggestion && (
-        <div className={style.suggestionFeedback}>
-          Mente du{' '}
-          <button
-            type="button"
-            className={style.inlineButton}
-            onClick={() => handleSuggestionClick(fetcher.data?.suggestion)}
-          >
-            {fetcher.data?.suggestion}
-          </button>
-          ?
-        </div>
-      )}
-    </>
-  );
-}
