+++ app/styles/ny-term.module.css
@@ -18,20 +18,6 @@
   margin-right: 10px;
 }
 
-.inlineButton {
-  border-style: none;
-  background: none;
-  color: inherit;
-  padding: 0;
-  text-decoration: underline;
-}
-
-.suggestionFeedback {
-  margin-top: 0.25rem;
-  font-size: 0.875em;
-  filter: brightness(85%);
-}
-
 :global(.form-control.is-invalid) {
   background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 12 12' width='12' height='12' fill='none' stroke='%23e57316'%3e%3ccircle cx='6' cy='6' r='4.5'/%3e%3cpath stroke-linejoin='round' d='M5.8 3.6h.4L6 6.5z'/%3e%3ccircle cx='6' cy='8.2' r='.6' fill='%23e57316' stroke='none'/%3e%3c/svg%3e");
 }
